add_items_widgets_splitted
